Added the interactive module
Interactive allows the user to play a game of exasott in the shell. The
exasott.__main__ module just calls interactive.main, so its possible to
start a game from the command line with:

    $ python -m exasott

Only works with py3k at the moment. I'll branch a py2k version later, if
the need arises.